refactor(cockpit): extract proof evidence input DTOs
